Remove everything that was obsolete
