+++ CupCake/CupCake.csproj
@@ -54,14 +54,13 @@
     <Compile Include="CupCakeMuffin.cs" />
     <Compile Include="Host\CupCakeComponents.cs" />
     <Compile Include="Host\CupCakeClient.cs" />
-    <Compile Include="Command.cs" />
     <Compile Include="CupCakeManualMuffin.cs" />
     <Compile Include="CupCakeMuffinPart.cs" />
     <Compile Include="Host\CupCakeClientArgs.cs" />
     <Compile Include="Host\EnumHelper.cs" />
     <Compile Include="ICommand.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="RelayCommand.cs" />
+    <Compile Include="CommandHandle.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\CupCake.Actions\CupCake.Actions.csproj">
